Grupo 11 - Requisito 8 funcionando, mas com detalhes a serem arrumados (pendente)
