+++ src/components/Product.jsx
@@ -4,12 +4,23 @@ import { Link } from 'react-router-dom';
 
 class Product extends Component {
   render() {
-    const { title, thumbnail, price, id } = this.props;
+    const { title, thumbnail, price, id, addCart } = this.props;
     return (
       <div data-testid="product">
         <img src={ thumbnail } alt={ title } />
         <h3>{title}</h3>
         <p>{`R$ ${price}`}</p>
+        <button
+          data-testid="product-add-to-cart"
+          type="button"
+          onClick={ addCart }
+          imagem={ thumbnail }
+          name={ title }
+          value={ price }
+          id={ id }
+        >
+          ADICIONAR AO CARRINHO
+        </button>
         <Link
           to={ {
             pathname: `/details/${id}`,
